+++ src/fides/api/ops/api/v1/endpoints/dataset_endpoints.py
@@ -338,6 +338,8 @@ def create_or_update_dataset(
 ) -> None:
     try:
         if connection_config.connection_type == ConnectionType.saas:
+            # Validating here instead of on ctl_dataset creation because this only applies
+            # when a ctl_dataset is being linked to a Saas Connector.
             _validate_saas_dataset(connection_config, dataset)  # type: ignore
         # Try to find an existing DatasetConfig matching the given connection & key
         dataset_config = create_method(db, data=data)