+++ tests/ops/fixtures/postgres_fixtures.py
@@ -49,7 +49,7 @@ def postgres_example_test_dataset_config(
         data={
             "connection_config_id": connection_config.id,
             "fides_key": fides_key,
-            "dataset": postgres_dataset,  # TODO Unified Fides Resources - Stop writing to this field
+            "dataset": postgres_dataset,
             "ctl_dataset_id": ctl_dataset.id,
         },
     )